Delete: 필요없는 파일들 삭제
user-controller service 다 삭제
