+++ src/controllers/index.js
@@ -1,6 +1,5 @@
-import userController from './user-controller';
+import authController from './auth-controller';
 import productController from './product-controller';
 import orderController from './order-controller';
-import authController from './auth-controller';
 
-export { userController, productController, orderController, authController };
+export { authController, productController, orderController };
