+++ src/services/index.js
@@ -1,7 +1,6 @@
-import userService from './user-service';
 import orderService from './order-service';
 import authService from './auth-service';
 import viewService from './view-service';
 import productService from './product-service';
 
-export { userService, orderService, authService, viewService, productService };
+export { authService, productService, orderService, viewService };
