--- src/services/user-service.js
@@ -1,7 +1,0 @@
-import { userModel } from '../db/index.js';
-import { customError } from '../middlewares/index.js';
-import { authService } from './index.js';
-
-const userService = {};
-
-export default userService;
